update articles.views with login_required decoratiors and login redirection logic

+++ accounts/views.py
@@ -3,10 +3,11 @@ from django.contrib.auth.forms import AuthenticationForm, PasswordChangeForm
 from django.contrib.auth import login as auth_login
 from django.contrib.auth import logout as auth_logout
 from django.contrib.auth import update_session_auth_hash
+from django.views.decorators.http import require_http_methods, require_POST
 from .forms import CustomUserCreationForm, CustomUserChangeForm
 # Create your views here.
 
-
+@require_http_methods(['GET', 'POST'])
 def login(request):
     if request.user.is_authenticated:
         return redirect('articles:index')
@@ -15,18 +16,20 @@ def login(request):
             form = AuthenticationForm(request, request.POST)
             if form.is_valid():
                 auth_login(request, form.get_user())
-                return redirect('articles:index')
+                return redirect(request.GET.get('next') or 'articles:index')
         else:
             form = AuthenticationForm
         context = { 'form': form }
         return render(request, 'accounts/login.html', context)
 
-
+@require_POST
 def logout(request):
-    auth_logout(request)
-    return redirect('articles:index')
+    if request.user.is_authenticated:
+        auth_logout(request)
+        return redirect('articles:index')
 
 
+@require_http_methods(['GET', 'POST'])
 def signup(request):
     if request.method == 'POST':
         form = CustomUserCreationForm(request.POST)
@@ -39,33 +42,38 @@ def signup(request):
     context = {'form':form,}
     return render(request, 'accounts/signup.html', context)
 
-
+@require_POST
 def delete(request):
-    request.user.delete()
-    auth_logout(request)
-    return redirect('articles:index')
+    if request.user.is_authenticated:
+        request.user.delete()
+        auth_logout(request)
+        return redirect('articles:index')
 
 
+@require_http_methods(['GET', 'POST'])
 def update(request):
-    if request.method=="POST":
-        form = CustomUserChangeForm(request.POST, instance=request.user)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:index')
-    else:
-        form = CustomUserChangeForm(instance=request.user)
-    context = {'form':form}
-    return render(request, 'accounts/update.html', context)
+    if request.user.is_authenticated:
+        if request.method=="POST":
+            form = CustomUserChangeForm(request.POST, instance=request.user)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:index')
+        else:
+            form = CustomUserChangeForm(instance=request.user)
+        context = {'form':form}
+        return render(request, 'accounts/update.html', context)
 
 
+@require_http_methods(['GET', 'POST'])
 def change_password(request):
-    if request.method == 'POST':
-        form = PasswordChangeForm(request.user, request.POST)
-        if form.is_valid():
-            form.save()
-            update_session_auth_hash(request, form.user)
-            return redirect('articles:index')
-    else:
-        form = PasswordChangeForm(request.user)
-    context = { 'form':form }
-    return render(request, 'accounts/change_password.html', context)
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            form = PasswordChangeForm(request.user, request.POST)
+            if form.is_valid():
+                form.save()
+                update_session_auth_hash(request, form.user)
+                return redirect('articles:index')
+        else:
+            form = PasswordChangeForm(request.user)
+        context = { 'form':form }
+        return render(request, 'accounts/change_password.html', context)
