+++ articles/views.py
@@ -2,6 +2,9 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .models import Article
 from .forms import ArticleForm
 from django.views.decorators.http import require_http_methods, require_safe, require_POST
+from django.contrib.auth.decorators import login_required
+
+
 # Create your views here.
 @require_safe
 def index(request):
@@ -17,10 +20,13 @@ def detail(request, pk):
 
 @require_POST
 def delete(request, pk):
-    article = get_object_or_404(Article, pk=pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=pk)
+        article.delete()
     return redirect('articles:index')
 
+
+@login_required
 @require_http_methods(['GET', 'POST'])
 def create(request):
     if request.method == "POST":
@@ -33,6 +39,8 @@ def create(request):
     context = {'form':form}
     return render(request, 'articles/create.html', context)
 
+
+@login_required
 @require_http_methods(['GET', 'POST'])
 def update(request, pk):
     article = get_object_or_404(Article, pk=pk)
